add logic

+++ src/main/java/com/ilevitsky/testproject/tasksystem/service/impl/TaskServiceImpl.java
@@ -1,16 +1,26 @@
 package com.ilevitsky.testproject.tasksystem.service.impl;
 
+import com.ilevitsky.testproject.tasksystem.dto.CommentDto;
 import com.ilevitsky.testproject.tasksystem.dto.TaskDto;
+import com.ilevitsky.testproject.tasksystem.dto.paging.PageInfo;
+import com.ilevitsky.testproject.tasksystem.dto.paging.PagedResponse;
+import com.ilevitsky.testproject.tasksystem.entity.Task;
+import com.ilevitsky.testproject.tasksystem.entity.TaskStatus;
+import com.ilevitsky.testproject.tasksystem.exception.OperationDeniedException;
 import com.ilevitsky.testproject.tasksystem.exception.TaskNotFoundException;
 import com.ilevitsky.testproject.tasksystem.mapper.TaskMapper;
+import com.ilevitsky.testproject.tasksystem.mapper.UserMapper;
 import com.ilevitsky.testproject.tasksystem.repository.TaskRepository;
+import com.ilevitsky.testproject.tasksystem.repository.UserRepository;
 import com.ilevitsky.testproject.tasksystem.service.TaskService;
+import com.ilevitsky.testproject.tasksystem.util.AuthUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Objects;
 import java.util.UUID;
 
 @Slf4j
@@ -19,39 +29,87 @@ import java.util.UUID;
 public class TaskServiceImpl implements TaskService {
 
   private final TaskRepository taskRepository;
+  private final UserRepository userRepository;
   private final TaskMapper taskMapper;
+  private final UserMapper userMapper;
+  private final AuthUtil authUtil;
 
   @Override
-  public List<TaskDto> getAll(String status) {
-    if (Objects.isNull(status)) {
-      return taskRepository.findAll().stream().map(taskMapper::mapToDto).toList();
-    }
-    return taskRepository.findAll().stream()
-        .filter(task -> task.getStatus().getLowercaseName().equals(status))
-        .map(taskMapper::mapToDto)
-        .toList();
+  public PagedResponse<TaskDto> getAll(Pageable pageable) {
+    if (authUtil.isAdmin()) {
+      Page<Task> page = taskRepository.findAll(pageable);
+
+      return new PagedResponse<>(
+          page.getContent().stream().map(taskMapper::mapToDto).toList(),
+          new PageInfo(
+              page.getNumber(),
+              page.getSize(),
+              page.getTotalElements(),
+              page.getTotalPages(),
+              page.isLast()));
+    } else
+      throw new OperationDeniedException(
+          "Forbidden operation for current user! Must be admin to get all tasks.");
   }
 
   @Override
   public TaskDto getById(UUID id) {
-    return taskRepository
-        .findById(id)
-        .map(taskMapper::mapToDto)
-        .orElseThrow(() -> new TaskNotFoundException(String.format("No task with id %s", id)));
+    var task =
+        taskRepository
+            .findById(id)
+            .orElseThrow(() -> new TaskNotFoundException(String.format("No task with id %s", id)));
+    if (authUtil.isAdmin() || authUtil.isSameUser(task.getAssignee().getEmail())) {
+      return taskMapper.mapToDto(task);
+    } else
+      throw new OperationDeniedException(
+          "Forbidden operation for current user! Must be admin or task assignee to view task.");
   }
 
   @Override
   public TaskDto create(TaskDto dto) {
-    dto.setId(null);
-    return mapAndSave(dto);
+    if (authUtil.isAdmin()) {
+      dto.setId(null);
+      var creator =
+          userRepository
+              .findByEmail(authUtil.getCurrentUserEmail())
+              .orElseThrow(
+                  () ->
+                      new UsernameNotFoundException(
+                          String.format(
+                              "No admin with email %s, unable to create task.",
+                              dto.getCreator().getEmail())));
+      dto.setCreator(userMapper.mapToDto(creator));
+
+      var assignee = userRepository.findByEmail(dto.getAssignee().getEmail()).orElse(null);
+      dto.setAssignee(userMapper.mapToDto(assignee));
+
+      return mapAndSave(dto);
+    } else
+      throw new OperationDeniedException(
+          "Forbidden operation for current user! Must be admin to create tasks.");
   }
 
   @Override
   public TaskDto update(UUID id, TaskDto dto) {
-    if (!taskRepository.existsById(id)) {
-      throw new TaskNotFoundException(String.format("No task with id %s", id));
+    var task =
+        taskRepository
+            .findById(id)
+            .orElseThrow(() -> new TaskNotFoundException(String.format("No task with id %s", id)));
+
+    if (authUtil.isAdmin()) {
+      var newTask = taskMapper.createTaskEntityWithoutId(dto);
+      newTask.setId(id);
+      newTask.setCreator(task.getCreator());
+
+      var assignee = userRepository.findByEmail(dto.getAssignee().getEmail()).orElse(null);
+
+      newTask.setAssignee(assignee);
+      return taskMapper.mapToDto(taskRepository.save(newTask));
+    } else {
+      task.setStatus(TaskStatus.valueOf(dto.getStatus()));
+      task.setDescription(dto.getDescription());
     }
-    var task = taskMapper.createTaskEntityWithoutId(dto);
+
     task.setId(id);
 
     return taskMapper.mapToDto(taskRepository.save(task));
@@ -59,7 +117,29 @@ public class TaskServiceImpl implements TaskService {
 
   @Override
   public void delete(UUID id) {
-    taskRepository.deleteById(id);
+    if (authUtil.isAdmin()) {
+      taskRepository.deleteById(id);
+    } else
+      throw new OperationDeniedException(
+          "Forbidden operation for current user! Must be admin to delete tasks.");
+  }
+
+  @Override
+  public TaskDto addComment(UUID id, CommentDto comment) {
+    var task =
+        taskRepository
+            .findById(id)
+            .orElseThrow(() -> new TaskNotFoundException(String.format("No task with id %s", id)));
+
+    if (authUtil.isAdmin() || authUtil.isSameUser(task.getAssignee().getEmail())) {
+      var comments = task.getComments();
+      comments.add(comment.getComment());
+      task.setComments(comments);
+
+      return taskMapper.mapToDto(taskRepository.save(task));
+    } else
+      throw new OperationDeniedException(
+          "Forbidden operation for current user. Must be admin or task assignee to add comments.");
   }
 
   private TaskDto mapAndSave(TaskDto dto) {
